“fix”

+++ app/Http/Controllers/TasksController.php
@@ -15,8 +15,22 @@ class TasksController extends Controller
      */
     public function index()
     {
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
         // メッセージ一覧を取得
-        $tasks = Task::all();
+        //$tasks = Task::all();
 
         // メッセージ一覧ビューでそれを表示
         return view('tasks.index', [
@@ -142,8 +156,14 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         // メッセージを削除
-        $task->delete();
+        //$task->delete();
 
         // トップページへリダイレクトさせる
         return redirect('/');
